% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSFM_1D.R
\name{dataDSFM1D}
\alias{dataDSFM1D}
\title{DSFM Data Set Generation}
\usage{
dataDSFM1D(y, x1 = NULL)
}
\arguments{
\item{y}{a numeric matrix of the data. The first column is a time indicator
with a proper \code{date} format. The remainings columns are the responses
\code{y} for each covariates.}

\item{x1}{a numeric vector of the covariates.}
}
\value{
\code{dataDSFM1D} returns a list, which belongs to the class
\code{"DSFM1DData"}. The list contains the dates, the responses, and
the covariates in three distinct columns.

The generic functions \code{print},\code{summary}, and \code{plot}
are available for this class.
}
\description{
This function generates a data set to be used in the \code{\link{DSFM}}
function.
}
\note{
A data set with the correct \code{DSFM1D} class format can be
inserted as \code{y} to return an \code{object} of class \code{"DSFM1DData"}.
}
\seealso{
\code{\link{summary.DSFM1DData}}, \code{\link{plot.DSFM1DData}}.
}

