% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSFM_2D.R
\name{dataDSFM2D}
\alias{dataDSFM2D}
\title{DSFM Data Set Generation}
\usage{
dataDSFM2D(y, x1 = NULL, x2 = NULL)
}
\arguments{
\item{y}{a numeric list of matrix containing the response variables.
Each matrix represents one date.}

\item{x1}{a numeric vector of the covariates x1.}

\item{x2}{a numeric vector of the covariates x2.}
}
\value{
\code{dataDSFM2D} returns a list, which belongs to the class
\code{"DSFM2DData"}. The list contains the dates, the responses and
the covariates in four distinct columns.

The generic functions \code{print},\code{summary}, and \code{plot}
are available for this class.
}
\description{
This function generates a data set to be used in the \code{\link{DSFM}}
function.
}
\note{
If the data set contains different covariates for each dates, the data
has to be formatted manually then inserted in the \code{dataDSFM2D}
function to obtain an object of class \code{DSFM2DData}.

The correct format is a four columns \code{data.frame} containing a time
indicator as first columns, the responses in the second columns and the
coordinates of the two dimensions in the two remaining columns.
}
\seealso{
\code{\link{summary.DSFM2DData}}, \code{\link{plot.DSFM2DData}}.
}

