% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSFM.R
\name{DSFM}
\alias{DSFM}
\title{Estimation Using Dynamic Semiparametric Factor Model}
\usage{
DSFM(data, numDataPoints = 25, h = 0.05, L = 3, initialLoad = "WN",
  tol = 1e-05, maxIt = 301)
}
\arguments{
\item{data}{a matrix containing time indicator in first row, value
\eqn{Y_{t,j}} in second row, and the coordinates \eqn{X_{t,j}} in the
remaining rows. Proper formatting has to be done using the
 \code{\link{DSFM1DData}} or \code{\link{DSFM2DData}} functions.}

\item{numDataPoints}{the number of points in the axis of the grid to perform
the kernel density estimation.}

\item{h}{the bandwidth used to perform the kernel density estimation. In one
dimension, can be either a single global parameter, or a vector of the same
length of numDataPoints to perform local kernel estimation. In two dimension,
can be a single global parameter, a vector of two bandwidths (one by
dimension) or a matrix of size \eqn{numDataPoints x 2} for local bandwidth.}

\item{L}{the number of underlying factors.}

\item{initialLoad}{the type of initial loadings to choose between White Noise
\code{"WN"}, and AR(1) process \code{"AR"}. Required as starting value of the
algorithm. Changing the \code{initialLoad} can sligthly improve the
convergence rate.}

\item{tol}{the tolerance for the algorithm to stop.}

\item{maxIt}{the maximum number of iterations for the algorithm to break.}
}
\value{
\code{DSFM} returns an object of class \code{"DSFM1D"} or
\code{"DSFM2D"} depending on the dimension of the input data.

The generic functions \code{print}, \code{summary}, \code{plot} and
\code{predict} are available.

An object of class \code{"DSFM1D"} is a list containing the following
components:
\item{\code{Data}}{the input data.}
\item{\code{Y}}{the input data in a more usual format, i.e. a matrix with a
time indicator as first row and the following rows being the value
\eqn{Y_{t,j}} for each covariates \eqn{X_{t,j}}.}
\item{\code{YHat}}{the estimated \eqn{\hat{Y}_{t,j}} with the same format,
i.e. a matrix with a time indicator as first row and the following rows being
the value \eqn{\hat{Y}_{t,j}} for each covariates \eqn{X_{t,j}}.}
\item{\code{ZHat}}{the estimated factor loadings \eqn{\hat{Z}_{t,j}}.}
\item{\code{mHat}}{the estimated factor functions \eqn{\hat{m}_l}.}
\item{\code{EV}}{gives the Explained Variance, used to select the approriate
number of factors.}
\item{\code{RMSE}}{gives the Root Mean Squared Error, used to compare the
goodness-of-fit between models.}
\item{\code{Bw}}{gives the bandwidth \eqn{h} used and two selection criteria to
select the optimal bandwidth.}
\item{\code{x1}}{the vector of the covariates.}
\item{\code{Density}}{the kernel density estimation performed.}
\item{\code{Convergence}}{the value of the algorithm stopping criterion at
each loop.}
\item{\code{Time}}{an indicator of the time taken by the function to perform
the fit.}

Object of class \code{"DSFM2D"} provides the same outputs except that the
\code{Y} and \code{YHat} are kept in the specific format used by the function.
}
\description{
\code{DSFM} performs a model estimation using Dynamic Semiparametric Factor
mechanics.
}
\details{
Dynamic Semiparametric Factor Models (DSFM) are defined as
\eqn{ Y_{t,j} = m_0(X_{t,j}) + \sum_{l=1}^L Z_{t,l} m_l(X_{t,j}) +
\varepsilon_{t,j}}.
DSFM estimation is performed using kernel density for the non-parametric
functions \eqn{m_l}. The estimation is performed using  the iterative
algorithm of Fengler and al. (2007).

The function has predefined arguments that can be changed for better
approximation.
First, the number of data points on the estimation grid is set
to 25. Larger grid can significantly increase the computation time without
necesseraly improve the fit.
Secondly, the bandwidth \eqn{h} is basically set to 0.05 but optimal bandwidth
has to be found externally. The algorithm always normalize the covariates
to work on an estimation grid bounded beetween [0,1].

For model selection, different criteria are computed.

For number of factors selection, the function compute the Explained Variance,
for bandwidth selection, two criteria are computed, a weighted Akaike
Information Criterion (AIC) and a weighted Schwarz Criterion (SC).
The goodness-of-fit is measured by the Root Mean Squared Error (RMSE). The
proper definition of each criterion can be found in references.
}
\examples{

## One-Dimensional Data #################################################### #

# Prepare the data --------------------------------------------------------- #
# Interest rate of zero-coupon bond yield curves. Data from Bank of Canada.
data(canadianYieldCurves)
maturity <- c(1/4, 1/2, 3/4, 1:10, 20, 30)
dsfmData <- DSFM1DData(canadianYieldCurves[1:100, ], maturity)
dsfmData
plot(dsfmData)

# Set the parameters ------------------------------------------------------- #
h        <- 0.167
L        <- 3

# Fit the model ------------------------------------------------------------ #
dsfmFit  <- DSFM(dsfmData, h = h, L = L)
summary(dsfmFit)
plot(dsfmFit)

# Perform prediction ------------------------------------------------------- #
horizon  <- 5
predict(dsfmFit, nAhead = horizon)



## Two-Dimensional Data #################################################### #

# Prepare the data --------------------------------------------------------- #
simulatedData <- simulateDSFM2D()
dsfmData      <- simulatedData$dataSim

# Set the parameters ------------------------------------------------------- #
h        <- c(0.05, 0.05)
L        <- 3

# Fit the model ------------------------------------------------------------ #
dsfmFit  <- DSFM(dsfmData, h = h, L = L)
summary(dsfmFit)
plot(dsfmFit)

# Perform prediction ------------------------------------------------------- #
horizon  <- 5
predict(dsfmFit, nAhead = horizon)

}
\author{
The implementation of model by Marc Gumowski was based on
Fengler and al. (2007).
}
\references{
Borak, Szymon, Matthias R. Fengler, and Wolfgang K. Haerdle (2005).
"DSFM Fitting of Implied Volatility Surfaces". In: \emph{5th International
Conference on Intelligent Systems Design and Applications (ISDA'05)},
pp. 526-531. IEEE.

Fengler, Matthias R, Wolfgang K Haerdle, and Enno Mammen (2007).
"A Semiparametric Factor Model for Implied Volatility Surface Dynamics".
In: \emph{Journal of Financial Econometrics 5.2}, pp. 189-218.

Haerdle, Wolfgang K., and Piotr Majer (2014).
"Yield Curve Modeling and Forecasting using Semiparametric Factor Dynamics".
In: \emph{The European Journal of Finance}, pp. 1-21.
}
\seealso{
\code{\link{summary.DSFM1D}} / \code{\link{summary.DSFM2D}} for
summaries and \code{\link{plot.DSFM1D}} / \code{\link{plot.DSFM2D}} for plot
possibilities.

\code{\link{predict.DSFM1D}} / \code{\link{predict.DSFM2D}} provide succint
predictions.

\code{\link{DSFM1DData}} / \code{\link{DSFM2DData}} have to be used before
using the \code{\link{DSFM}} function to ensure that the data are correctly
formated.

\code{\link{simulateDSFM1D}} / \code{\link{simulateDSFM2D}} are functions
to simulate data that can be used as simple example purposes.
}

