% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSFM_2D.R
\name{DSFM2D}
\alias{DSFM2D}
\title{Estimation of Dynamic Semiparametric Factor Model for Two-Dimensional Data}
\usage{
DSFM2D(data, numDataPoints = 25, h = c(0.5, 0.5), L = 3,
  initialLoad = "WN", tol = 1e-05, maxIt = 301)
}
\arguments{
\item{data}{a matrix containing time indicator in first row, value
\eqn{Y_{t,j}} in second row, and the coordinates \eqn{X_{t,j}} in the two
remaining rows. Proper formatting has to be done using the
\code{\link{dataDSFM2D}}.}

\item{numDataPoints}{the number of points in the axis of the grid to perform
the kernel density estimation.}

\item{h}{the bandwidth used to perform the kernel density estimation. can be
a single global parameter, a vector of two bandwidths (one by dimension) or a
matrix of size \eqn{numDataPoints x 2} for local bandwidth.}

\item{L}{the number of underlying factors.}

\item{initialLoad}{the type of initial loadings to choose between White Noise
\code{"WN"}, and AR(1) process \code{"AR"}. Required as starting value of the
algorithm. Changing the \code{initialLoad} can sligthly improve the
convergence rate.}

\item{tol}{the tolerance for the algorithm to stop.}

\item{maxIt}{the maximum number of iterations for the algorithm to break.}
}
\value{
\code{DSFM2D} returns an object of class \code{"DSFM2D"}.

The generic functions \code{print},\code{summary}, \code{plot} and
\code{predict} are available.

An object of class \code{"DSFM2D"} is a list containing the following
components:
\item{\code{Data}}{the input data.}
\item{\code{YHat}}{the estimated \eqn{\hat{Y}_{t,j}}.}
\item{\code{ZHat}}{the estimated factor loadings \eqn{\hat{Z}_{t,j}}.}
\item{\code{mHat}}{the estimated factor functions \eqn{\hat{m}_l}.}
\item{\code{EV}}{gives the Explained Variance, used to select the approriate
number of factors.}
\item{\code{RMSE}}{gives the Root Mean Squared Error, used to compare models.}
\item{\code{Bw}}{gives the bandwidth \eqn{h} used and two selection criteria
 to select the optimal bandwidth.}
\item{\code{Density}}{the kernel density estimation performed.}
\item{\code{Convergence}}{the value of the algorithm stopping criterion at
each loop.}
\item{\code{Time}}{an indicator of the time taken by the function to perform
the fit.}
}
\description{
\code{DSFM2D} performs a model estimation using Dynamic Semiparametric Factor
mechanics with two-dimensional covariates. This function is called by the
\code{\link{DSFM}} main function.
}
\details{
Dynamic Semiparametric Factor Models (DSFM) are defined as
\eqn{ Y_{t,j} = m_0(X_{t,j}) + \sum_{l=1}^L Z_{t,l} m_l(X_{t,j}) +
\varepsilon_{t,j}}.
DSFM estimation is performed using kernel density for the non-parametric
functions \eqn{m_l}. The estimation is performed using  the iterative
algorithm of Fengler and al. (2007).

The function has predefined arguments that can be changed for better
approximation.
First, the number of data points on the estimation grid is set
to 25. Larger grid can significantly increase the computation time without
necesseraly improve the fit.
Secondly, the bandwidth \eqn{h} is basically set to 0.05 but optimal
bandwidth has to be found externally. The algorithm always normalize the
covariates to work on an estimation grid bounded beetween [0,1].

For model selection, different criteria are computed.

For number of factors selection, the function compute the Explained Variance,
for bandwidth selection, two criteria are computed, a weighted Akaike
Information Criterion (AIC) and a weighted Schwarz Criterion (SC).
The goodness-of-fit is measured by the Root Mean Squared Error (RMSE). The
proper definition of each criterion can be found in references.
}
\examples{
# Prepare the data --------------------------------------------------------- #
simulatedData <- simulateDSFM2D()
dsfmData      <- simulatedData$dataSim

# Set the parameters ------------------------------------------------------- #
h        <- c(0.05,0.05)
L        <- 3

# Fit the model ------------------------------------------------------------ #
dsfmFit  <- DSFM(dsfmData, h = h, L = L)
summary(dsfmFit)
plot(dsfmFit)

# Perform prediction ------------------------------------------------------- #
horizon  <- 5
predict(dsfmFit, nAhead = horizon)

}
\author{
The implementation of model by Marc Gumowski was based on
Fengler and al. (2007).
}
\references{
Fengler, Matthias R, Wolfgang K Haerdle, and Enno Mammen (2007).
"A Semiparametric Factor Model for Implied Volatility Surface Dynamics".
In: \emph{Journal of Financial Econometrics 5.2}, pp. 189-218.

Haerdle, Wolfgang K., and Piotr Majer (2014).
"Yield Curve Modeling and Forecasting using Semiparametric Factor Dynamics".
In: \emph{The European Journal of Finance}, pp. 1-21.

Borak, Szymon, Matthias R. Fengler, and Wolfgang K. Haerdle (2005)."DSFM
Fitting of Implied Volatility Surfaces". In: \emph{5th International
Conference on Intelligent Systems Design and Applications (ISDA'05)},
pp. 526-531. IEEE.
}
\seealso{
\code{\link{summary.DSFM2D}} for
summaries and \code{\link{plot.DSFM2D}} for plot
possibilities.

\code{\link{predict.DSFM2D}} provide succint
predictions.

\code{\link{dataDSFM2D}} has to be used before
using the \code{\link{DSFM}} function to ensure that the data are correctly
formated.

\code{\link{simulateDSFM2D}} is a function to simulate data that can be used
as simple example purposes.
}

