% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dysefamor.R
\docType{package}
\name{dysefamor}
\alias{dysefamor}
\alias{dysefamor-package}
\title{dysefamor: Estimation Algorithm of Dynamic Semiparametric Factor Model}
\description{
The dysefamor package provides two categories of important functions:
DSFM for one-dimensional data and DSFM for two-dimensional data.
}
\section{DSFM1D functions}{

\itemize{
\item{\code{\link{dataDSFM1D}}}
\item{\code{\link{summary.DSFM1DData}}}
\item{\code{\link{plot.DSFM1DData}}}
\item{\code{\link{DSFM1D}}}
\item{\code{\link{summary.DSFM1D}}}
\item{\code{\link{plot.DSFM1D}}}
\item{\code{\link{predict.DSFM1D}}}
}
}

\section{DSFM2D functions}{

\itemize{
\item{\code{\link{dataDSFM2D}}}
\item{\code{\link{summary.DSFM2DData}}}
\item{\code{\link{plot.DSFM2DData}}}
\item{\code{\link{DSFM2D}}}
\item{\code{\link{summary.DSFM2D}}}
\item{\code{\link{plot.DSFM2D}}}
\item{\code{\link{predict.DSFM2D}}}
}
}

