% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSFM.R
\name{\%^\%}
\alias{\%^\%}
\title{Power of matrix operator}
\usage{
x \%^\% n
}
\arguments{
\item{x}{a matrix.}

\item{n}{the power.}
}
\description{
\code{\%^\%} computes the matrix x at the power n. Inner function of the
DSFM iterative algorithm.
}

