% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSFM.R
\name{KernelDensity2D}
\alias{KernelDensity2D}
\title{Two-Dimensional Kernel Estimation}
\usage{
KernelDensity2D(y, I, J, x1, x2, u, U, h)
}
\arguments{
\item{y}{a list of numeric matrix of the data at each time \eqn{t}.}

\item{I}{the total number of matrix in \code{y}.}

\item{J}{a list of the total number of covariates in each matrix of \code{y}.}

\item{x1}{a list of numeric vector for the first dimension at each time
\eqn{t}.}

\item{x2}{a list of numeric vector for the second dimension at each time
\eqn{t}.}

\item{u}{a numeric matrix of the estimation grid.}

\item{U}{the length of \code{u}.}

\item{h}{a numeric matrix of the bandwidth.}
}
\value{
It returns \eqn{\hat{p}_t(u)}, \eqn{\hat{q}_t(u)} and the couples
\eqn{J_t\hat{p}_t(u)} and \eqn{J_t\hat{q}_t(u)}.
}
\description{
This function computes the two-dimensional kernel estimation needed for the
\code{\link{DSFM2D}} funtion to work. Following Fengler and al. (2007), it
is using the product of two \code{\link{QuarticKernel1D}} functions.
}
\references{
Fengler, Matthias R, Wolfgang K Haerdle, and Enno Mammen (2007).
"A Semiparametric Factor Model for Implied Volatility Surface Dynamics".
In: \emph{Journal of Financial Econometrics 5.2}, pp. 189-218.

Borak, Szymon, Matthias R. Fengler, and Wolfgang K. Haerdle (2005)."DSFM
Fitting of Implied Volatility Surfaces". In: \emph{5th International
Conference on Intelligent Systems Design and Applications (ISDA'05)},
pp. 526-531. IEEE.
}
\seealso{
Other kernel.functions: \code{\link{KernelDensity1D}},
  \code{\link{NormalKernel1D}},
  \code{\link{QuarticKernel1D}}
}

