% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSFM.R
\name{NormalKernel1D}
\alias{NormalKernel1D}
\title{Gaussian Kernel}
\usage{
NormalKernel1D(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\description{
This function computes a one-dimensional gaussian kernel.
}
\seealso{
Other kernel.functions: \code{\link{KernelDensity1D}},
  \code{\link{KernelDensity2D}},
  \code{\link{QuarticKernel1D}}
}

