% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSFM_1D.R
\name{plot.DSFM1D}
\alias{plot.DSFM1D}
\title{Plot Diagnostics for a DSFM1D Object}
\usage{
\method{plot}{DSFM1D}(x, which = "all", ask = TRUE, pal = "pink",
  col = "#016C59", type = "l", theta = 40, border = NA, box = T,
  shade = 0.2, expand = 0.5, ticktype = "detailed", ...)
}
\arguments{
\item{x}{an \code{object} of class \code{"DSFM1D"}.}

\item{which}{to choose between \code{"all"},\code{"loadings"}.
\code{"functions"},\code{"convergence"}, and \code{"fit"}.}

\item{ask}{logical; if TRUE, the user is asked before each plot, see
\code{\link{par}}(ask=.).}

\item{pal}{the color palette for the fit plot. To choose between
\code{"pink"},\code{"blue"},\code{"light"},\code{"dark"}.}

\item{col}{what color should be drawn.}

\item{type}{what type of plot should be drawn:
see \code{\link{plot.default}}.}

\item{theta}{angles defining the viewing direction of the fit plot.}

\item{border}{the color of the line drawn around the surface facets of the
fit plot. The default, NULL, corresponds to par("fg"). A value of NA will
disable the drawing of borders: this is sometimes useful when the surface is
shaded.}

\item{box}{should the bounding box for the surface be displayed.
The default is \code{FALSE}.}

\item{shade}{the shade at a surface facet is computed as ((1+d)/2)^shade,
where d is the dot product of a unit vector normal to the facet and a unit
vector in the direction of a light source. Values of shade close to one
yield shading similar to a point light source model and values close to zero
produce no shading. Values in the range 0.5 to 0.75 provide an approximation
to daylight illumination.}

\item{expand}{a expansion factor applied to the z coordinates. Often used
with 0 < expand < 1 to shrink the plotting box in the z direction.}

\item{ticktype}{character: "simple" draws just an arrow parallel to the axis
to indicate direction of increase; "detailed" draws normal ticks as per 2D
plots.}

\item{...}{other parameters to be passed through to plotting functions.}
}
\description{
Four plots (selectable by \code{which}) are currently available: a plot of
the convergence rate of the algorithm, a plot of the time series of the
estimated loadings for each factor, a plot of the estimated factor functions,
and a plot of the fit. By default, the four plots are provided.
}
\seealso{
\code{\link{dataDSFM1D}}, \code{\link{DSFM}}, \code{\link{DSFM1D}},
\code{\link{summary.DSFM1D}}, \code{\link{predict.DSFM1D}}.
}

