% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSFM_1D.R
\name{predict.DSFM1D}
\alias{predict.DSFM1D}
\title{Predict Method for One-Dimensional DSFM Fits}
\usage{
\method{predict}{DSFM1D}(object, nAhead = 12, p = 1, ...)
}
\arguments{
\item{object}{an object of class \code{"DSFM1D"}.}

\item{nAhead}{the number of steps ahead for which prediction is required.}

\item{p}{the order of the Vector Autoregressive Process to be fitted.}

\item{...}{other parameters to be passed through the \code{\link{VAR}}
and \code{\link{predict.varest}} functions.}
}
\value{
\code{predict.DSFM1D} returns an object of class
\code{"predict.DSFM1D"}. This class is a list containing:
\item{\code{YHatForecast}}{the forecasted responses.}
\item{\code{YHatForecastMatrix}}{the forecasted responses in a more usual
format.}
\item{\code{ZHatForecast}}{the predicted factors loadings.}
\item{\code{nAhead}}{the number of steps ahead.}
}
\description{
Predicted values based on DSFM object using Vector Autoregressive Processes
(VAR(p)).
}
\details{
This function makes uses of package \code{vars} to fit and predict VAR(p)
processes, before pluging in the predicted factor loadings in the DSFM. The
factors functions are interpolated by \code{\link{approx}}.
}
\references{
Bernhard Pfaff (2008). VAR, SVAR and SVEC Models: Implementation
Within R Package vars. In: \emph{Journal of Statistical Software 27(4)}.
URL http://www.jstatsoft.org/v27/i04/.
}
\seealso{
\code{\link{VAR}},\code{\link{predict.varest}}, \code{\link{DSFM}},
\code{\link{DSFM1D}},\code{\link{DSFM1DData}}.
}

