% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSFM_2D.R
\name{predict.DSFM2D}
\alias{predict.DSFM2D}
\title{Predict Method for Two-Dimensional DSFM Fits}
\usage{
\method{predict}{DSFM2D}(object, nAhead = 12,
  x1Forecast = unique(object$mHat[, 1]), x2Forecast = unique(object$mHat[,
  2]), p = 1, ...)
}
\arguments{
\item{object}{an object of class \code{"DSFM1D"}.}

\item{nAhead}{the number of steps ahead for which prediction is required.}

\item{x1Forecast}{the vector of covariates in the first dimension to be
forecasted.}

\item{x2Forecast}{the vector of covariates in the second dimension to be
forecasted.}

\item{p}{the order of the Vector Autoregressive Process to be fitted.}

\item{...}{other parameters to be passed through the \code{\link{VAR}}
and \code{\link{predict.varest}} functions.}
}
\value{
\code{predict.DSFM2D} returns an object of class
\code{"predict.DSFM2D"}. This class is a list containing:
\item{\code{YHatForecast}}{the forecasted responses.}
\item{\code{YHatForecastList}}{the forecasted responses in a more usual
format.}
\item{\code{ZHatForecast}}{the predicted factors loadings.}
\item{\code{nAhead}}{the number of steps ahead.}
}
\description{
Predicted values based on \code{"DSFM2D"} object using Vector Autoregressive
Processes (VAR(p)).
}
\details{
This function makes uses of package \code{vars} to fit and predict VAR(p)
processes, before pluging in the predicted factor loadings in the DSFM. The
factors functions are interpolated by the function \code{\link{interp}} of
package \code{akima}.
}
\references{
Bernhard Pfaff (2008). VAR, SVAR and SVEC Models: Implementation
Within R Package vars. In: \emph{Journal of Statistical Software 27(4)}.
URL http://www.jstatsoft.org/v27/i04/.

Hiroshi Akima and Albrecht Gebhardt (2015). akima: Interpolation of
Irregularly and Regularly Spaced Data. R package version 0.5-12.
URL http://CRAN.R-project.org/package=akima
}
\seealso{
\code{\link{VAR}},\code{\link{predict.varest}},
\code{\link{interp}}, \code{\link{DSFM}}, \code{\link{DSFM2D}},
\code{\link{DSFM2DData}}.
}

