% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSFM.R
\name{QuarticKernel1D}
\alias{QuarticKernel1D}
\title{Quartic Kernel}
\usage{
QuarticKernel1D(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\description{
This function computes a one-dimensional quartic kernel.
}
\seealso{
Other kernel.functions: \code{\link{KernelDensity1D}},
  \code{\link{KernelDensity2D}},
  \code{\link{NormalKernel1D}}
}

