% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSFM_1D.R
\name{simulateDSFM1D}
\alias{simulateDSFM1D}
\title{Simulate Responses for One-Dimensional Factor Models}
\usage{
simulateDSFM1D(model = "ns", n = 100, x1 = 1:30, L = 3, var = 5e-04,
  beta = c(0.065, -0.015, 0.05), tau = c(0.5, 6))
}
\arguments{
\item{model}{the type of model to be generated. To choose between \code{"ns"}
and \code{"dsfm"}.}

\item{n}{the number of observations.}

\item{x1}{a vector of covariates.}

\item{L}{the number of factors for the DSFM.}

\item{var}{the error \eqn{\varepsilon_{t,j}} of the models. Allows to control
the noise.}

\item{beta}{a vector of dimension (1x3) controlling the starting values of
the VAR(1) process.}

\item{tau}{a vector of dimension (1x2) specifying the parameters \eqn{\Tau_1}
and \eqn{\Tau_2} of the Extended Nelson-Siegel model. These parameters are
constant.}
}
\value{
\code{simulateDSFM1D} returns a list containing:
\item{\code{dataSim}}{an object of class \code{"DSFM1DData"}, output of the
\code{\link{DSFM1DData}} function. This object can be immediatly used by the
\code{\link{DSFM}} algorithm.}
\item{\code{YSim}}{the simulated data in a more usual format.}
\item{\code{Z_tl}}{the simulated factor loadings.}
\item{\code{x1}}{the vector of the covariates.}

Depending on the model simulated, the functions returns also:
\item{\code{m_l}}{the factors functions used to compute the DSFM.}
\item{\code{tau}}{the constant \eqn{\Tau_1} and \eqn{\Tau_2} used to compute
the Extended Nelson-Siegel model.}
}
\description{
This function simulates responses for one-dimensional factor models.
}
\details{
Two different way of generating data are available: using an Extended
Nelson-Siegel model following Bliss (1997) \code{"ns"} with a predefined
VAR(1) process, or a Dynamic Semiparametric Factor Model \code{"dsfm"}
with predefined factors functions.
This function is used for example purpose, only few parameters are available
to control the simulation.

The starting values to simulate the Extended
Nelson-Siegel model are taken following Linton and al. (2001).

The factors loadings of the DSFM are simulated from three independant AR(1)
process and the factors functions are predefined to be orthogonals.
}
\references{
Linton, Oliver et al. (2001). "Yield Curve Estimation by Kernel
Smoothing Methods". In: \emph{Journal of Econometrics 105.1}, pp. 185-223.

Bliss, Robert R.(1997). "Testing Term Structure Estimation Methods".
In: \emph{Advances in Futures and Options Research 9}, pp. 197-231.
}
\seealso{
\code{\link{DSFM1DData}}, \code{\link{DSFM}}, \code{\link{DSFM1D}}.
}

