% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSFM_2D.R
\name{simulateDSFM2D}
\alias{simulateDSFM2D}
\title{Simulate Responses for Two-Dimensional DSFM}
\usage{
simulateDSFM2D(n = 10, x1 = 1:25/25, x2 = 1:25/25, L = 3, var = 0.05)
}
\arguments{
\item{n}{the number of observations.}

\item{x1}{a vector of the first dimension of covariates. They will be constant
through time}

\item{x2}{a vector of the second dimension of covariates. They will be constant
through time}

\item{L}{the number of factors for the DSFM.}

\item{var}{the error \eqn{\varepsilon_{t,j}} of the models. Allows to control
the noise.}
}
\value{
\code{simulateDSFM2D} returns a list containing:
\item{\code{dataSim}}{an object of class \code{"DSFM2DData"}, output of the
\code{\link{dataDSFM2D}} function. This object can be immediatly used by the
\code{\link{DSFM}} algorithm.}
\item{\code{YSim}}{the simulated data in a more usual format.}
\item{\code{Z_tl}}{the simulated factor loadings.}
\item{\code{m_l}}{the factors functions used to compute the DSFM.}
}
\description{
This function simulates responses for two-dimensional DSFM.
}
\details{
This function is used for example purpose, only few parameters are available
to control the simulation. The factors loadings are generated using three
independant AR(1) process and the factors functions are predefined to be
orthogonals.
}
\seealso{
\code{\link{dataDSFM2D}}, \code{\link{DSFM}}, \code{\link{DSFM2D}}.
}

