% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSFM.R
\name{skewness}
\alias{skewness}
\title{Skewness}
\usage{
skewness(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector, matrix or data frame.}

\item{na.rm}{logical. Should missing values be removed?}
}
\description{
This function computes the skewness of given data.
}
\author{
Lukasz Komsta
}
\references{
Lukasz Komsta and Frederick Novomestky (2015). moments: Moments,
cumulants, skewness, kurtosis and related tests. R package
version 0.14. http://CRAN.R-project.org/package=moments
}

