% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSFM_1D.R
\name{summary.DSFM1DData}
\alias{summary.DSFM1DData}
\title{Summarizing One-Dimensional DSFM Data Set}
\usage{
\method{summary}{DSFM1DData}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"DSFM1DData"}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The function \code{summary.DSFM1DData} returns a summary
statistics of the data set given in \code{object}.
}
\description{
\code{summary} method for class \code{"DSFM1DData"}.
}
\seealso{
\code{\link{dataDSFM1D}}, \code{\link{plot.DSFM1DData}}.
}

