% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSFM_2D.R
\name{summary.DSFM2D}
\alias{summary.DSFM2D}
\title{Summarizing DSFM Fits for Two-Dimensional Data}
\usage{
\method{summary}{DSFM2D}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"DSFM2D"}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The function \code{summary.DSFM2D} returns a list of summary
statistics of the fitted DSFM given in \code{object}, using the components
(list elements) \code{"call"} from its arguments, plus:
\item{\code{ZHat}}{a summary of the estimated factor loadings.}
\item{\code{EV}}{the Explained Variance, used to select the approriate
number of factors.}
\item{\code{RMSE}}{the Root Mean Squared Error, used to compare the goodness-
of-fit between models.}
\item{\code{Bw}}{The bandwidth and its selection criteria.}
}
\description{
\code{summary} method for class \code{"DSFM2D"}.
}
\seealso{
\code{\link{dataDSFM2D}}, \code{\link{DSFM}}, \code{\link{DSFM2D}},
\code{\link{plot.DSFM2D}}, \code{\link{predict.DSFM2D}}.
}

