% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSFM_2D.R
\name{summary.DSFM2DData}
\alias{summary.DSFM2DData}
\title{Summarizing Two-Dimensional DSFM Data Set}
\usage{
\method{summary}{DSFM2DData}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"DSFM2DData"}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The function \code{summary.DSFM2DData} returns a summary
statistics of the data set given in \code{object} for each covariates
directions.
}
\description{
\code{summary} method for class \code{"DSFM2DData"}.
}
\seealso{
\code{\link{dataDSFM2D}}, \code{\link{plot.DSFM2DData}}.
}

